Cover trace provider and template edge cases
Move trace-provider internal tests into a sibling support module so the
provider stays within the repository size guideline while preserving access
to replay internals.

Exercise the real `TraceLlm` hint-mismatch overflow path, strengthen the
template expansion property test with a growing cycle that reaches the cap,
and preserve extracted JSON nulls during template substitution.

Document the shared TraceLlm fixture helpers used by support unit tests.
